Scriptlocator uses Unity_Webgl defines
Split out LocalFileScriptLocator for use within ulox core as it does not use the builtin scripts
