+++ ulox-example/Assets/ulox-example/Scripts/Library/ScriptLocator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aeb43dd24e396c34b8fe1b6a130d0c5f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
