+++ ulox/ulox.core/Package/Runtime/Engine/ScriptLocator.cs
@@ -1,39 +1,23 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 
 namespace ULox
 {
-    public class ScriptLocator : IScriptLocator
+    public class LocalFileScriptLocator : IScriptLocator
     {
-        private readonly Dictionary<string, string> _builtinScripts;
         private readonly DirectoryInfo _directory;
-
-        public ScriptLocator(
-            Dictionary<string, string> builtinScripts,
-            DirectoryInfo directory)
+        
+        public LocalFileScriptLocator()
         {
-            _builtinScripts = builtinScripts;
-            _directory = directory;
+            _directory = new DirectoryInfo(Environment.CurrentDirectory);
         }
-
-        public ScriptLocator()
-            : this(new Dictionary<string, string>(), new DirectoryInfo(Environment.CurrentDirectory))
-        {
-        }
-
-        public void Add(string name, string content)
-            => _builtinScripts[name] = content;
 
         public string Find(string name)
         {
-            if (_builtinScripts.TryGetValue(name, out var val))
-                return val;
-
             var externalMatches = Directory.GetFiles(_directory.FullName, $"{name}*");
             if (externalMatches != null && externalMatches.Length > 0)
                 return File.ReadAllText(externalMatches[0]);
-
+            
             return null;
         }
     }
