+++ ulox/ulox.core/Package/Runtime/Interfaces/IScriptLocator.cs
@@ -2,8 +2,6 @@
 {
     public interface IScriptLocator
     {
-        void Add(string name, string content);
-
         string Find(string name);
     }
 }
